swscale: partially move the arch specific code left
The swscale_template is split among ppc and x86, lots of code is
still duplicated and should be removed later.

Again uniform the init system to be more similar to the dsputil one

Unset h*scale_fast in the x86 init in order to make the output
consistent with the previous status. Thanks to Josh for spotting it.