dirac: introduce native decoder
This decoder is mainly result of the work of Marco Gerards <marco@gnu.org>
during the Summer of Code 2007. David Conrad <lessen42@gmail.com>
improved greatly the codebase during the following year and
Jordi Ortiz <nenjordi@gmail.com> completed it ensuring it is bitexact
and up to spec during the Summer of Code 2011.

Additional fixes and adaptations by Michael Niedermayer
<michaelni@gmx.at> and Luca Barbato <lu_zero@gentoo.org>

Signed-off-by: Luca Barbato <lu_zero@gentoo.org>